feat(game): Add game logic to seperate pkg
