+++ main.go
@@ -3,24 +3,18 @@ package main
 
 import (
 	"flag"
-	"fmt"
 	"go-snake/pkg/canvas"
+	"go-snake/pkg/game"
 	"go-snake/pkg/snake"
 	"log"
-	"time"
 
 	"github.com/gowhale/led-matrix-golang/pkg/config"
 	"github.com/gowhale/led-matrix-golang/pkg/gui"
-	term "github.com/nsf/termbox-go"
 )
 
 const (
 	size          = 8
 	defaultConfig = "eight-by-eight.json"
-)
-
-var (
-	errGameFin = fmt.Errorf("fin")
 )
 
 func main() {
@@ -56,90 +50,9 @@ func main() {
 
 	cvs := canvas.NewCanvas(size, size, &snk)
 
-	if err := gameLoop(scrn, cvs, &snk); err != nil && err != errGameFin {
+	if err := game.Loop(scrn, cvs, &snk); err != nil && err != game.ErrGameFin {
 		log.Panicln(err)
 	}
 	log.Println("GOOD GAME!")
 	log.Printf("Your score was %d", snk.Score())
-}
-
-func passTime(snk *snake.Snake, errs chan error) {
-	startTime := time.Now()
-	lastGrow := startTime
-	lastScreenRefresh := startTime
-	for {
-		timeSinceGrow := time.Since(lastGrow)
-		if timeSinceGrow > time.Second*5 {
-			lastGrow = time.Now()
-		}
-		timeSinceRefresh := time.Since(lastScreenRefresh)
-		if timeSinceRefresh > time.Millisecond*100 {
-			lastScreenRefresh = time.Now()
-			snk.NextMove()
-			if snk.Dead() {
-				errs <- errGameFin
-			}
-		}
-	}
-}
-
-func checkKeyInput(snk *snake.Snake, errs chan error) {
-	for {
-		switch ev := term.PollEvent(); ev.Type {
-		case term.EventKey:
-			switch ev.Key {
-			case term.KeyEsc:
-				errs <- fmt.Errorf("esc pressed")
-			case term.KeyArrowUp:
-				log.Println("Going North")
-				snk.ChangeDirection(snake.North)
-			case term.KeyArrowLeft:
-				log.Println("Going west")
-				snk.ChangeDirection(snake.West)
-			case term.KeyArrowRight:
-				log.Println("Going East")
-				snk.ChangeDirection(snake.East)
-			case term.KeyArrowDown:
-				log.Println("Going South")
-				snk.ChangeDirection(snake.South)
-			default:
-				log.Println("Invalid key")
-			}
-		case term.EventError:
-			errs <- ev.Err
-		}
-	}
-}
-func displayGame(scrn gui.Screen, cvs canvas.Canvas, errs chan error) {
-	for {
-		log.Println("Displaying!")
-		if err := scrn.DisplayMatrix(cvs.GetMatrix(), time.Millisecond*50); err != nil {
-			errs <- err
-		}
-		if err := scrn.AllLEDSOff(); err != nil {
-			errs <- err
-		}
-	}
-}
-
-func gameLoop(scrn gui.Screen, cvs canvas.Canvas, snk *snake.Snake) error {
-	if err := term.Init(); err != nil {
-		return err
-	}
-	defer term.Close()
-
-	errs := make(chan error, 1)
-
-	go passTime(snk, errs)
-
-	go checkKeyInput(snk, errs)
-
-	go displayGame(scrn, cvs, errs)
-
-	for {
-		err := <-errs
-		if err != nil {
-			return err
-		}
-	}
 }
